+++ src/Twig/TwigExtension.php
@@ -7,13 +7,8 @@
 
 namespace QL\Panthor\Twig;
 
-use DateTime;
-use DateTimeZone;
-use MCP\DataType\Time\Clock;
-use MCP\DataType\Time\TimePoint;
 use QL\Panthor\Utility\Url;
 use Twig_Extension;
-use Twig_SimpleFilter;
 use Twig_SimpleFunction;
 
 class TwigExtension extends Twig_Extension
@@ -24,35 +19,19 @@ class TwigExtension extends Twig_Extension
     private $url;
 
     /**
-     * @type Clock
-     */
-    private $clock;
-
-    /**
-     * @type string
-     */
-    private $displayTimezone;
-
-    /**
-     * @type boolean
+     * @type bool
      */
     private $isDebugMode;
 
     /**
      * @param Url $url
-     * @param Session $session
-     * @param CsrfManager $csrf
-     * @param Clock $clock
-     * @param string $timezone
-     * @param boolean $isDebugMode
+     * @param bool $isDebugMode
      */
-    public function __construct(Url $url, Clock $clock, $timezone, $isDebugMode)
+    public function __construct(Url $url, $isDebugMode)
     {
         $this->url = $url;
-        $this->clock = $clock;
 
-        $this->displayTimezone = $timezone;
-        $this->isDebugMode = $isDebugMode;
+        $this->isDebugMode = (bool) $isDebugMode;
     }
 
     /**
@@ -66,25 +45,13 @@ class TwigExtension extends Twig_Extension
     /**
      * {@inheritdoc}
      */
-    public function getFilters()
-    {
-        return [
-            new Twig_SimpleFilter('timepoint', [$this, 'formatTimePoint']),
-        ];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function getFunctions()
     {
         return [
             new Twig_SimpleFunction('urlFor', [$this->url, 'urlFor']),
             new Twig_SimpleFunction('route', [$this->url, 'currentRoute']),
 
-            new Twig_SimpleFunction('isDebugMode', [$this, 'isDebugMode']),
-
-            new Twig_SimpleFunction('timepoint', [$this, 'getTimepoint'])
+            new Twig_SimpleFunction('isDebugMode', [$this, 'isDebugMode'])
         ];
     }
 
@@ -94,44 +61,5 @@ class TwigExtension extends Twig_Extension
     public function isDebugMode()
     {
         return ($this->isDebugMode);
-    }
-
-    /**
-     * @param string|null $modifier
-     *
-     * @return TimePoint
-     */
-    public function getTimepoint($modifier = null)
-    {
-        $now = $this->clock->read();
-
-        if ($modifier) {
-            $now->modify($modifier);
-        }
-
-        return $now;
-    }
-
-    /**
-     * Format a DateTime or TimePoint. Invalid values will output an empty string.
-     *
-     * @param TimePoint|DateTime|null $time
-     * @param string $format
-     *
-     * @return string
-     */
-    public function formatTimepoint($time, $format)
-    {
-        if ($time instanceof TimePoint) {
-            return $time->format($format, $this->displayTimezone);
-        }
-
-        if ($time instanceof DateTime) {
-            $formatted = clone $time;
-            $formatted->setTimezone(new DateTimeZone($this->displayTimezone));
-            return $formatted->format($format);
-        }
-
-        return '';
     }
 }